IPC readbytes into a memorystream with timeOut
+ Json.net does not cope in some situations with long waits, leaving the
connection for both namedpipes and unixsockets hanging.
This can be workaround by parsing the json and timeout and closing the
json string (previous commit). This is very messy and is more efficient
to do it at lower level as before.

Parsing the Json directly from the stream it will improve some
performance.