Downgrade some projects to 4.6.1 & fix IpcClient build
Installing VS2017 community in a Microsoft
Surface laptop doesn't make it download
the .NET 4.6.2 by default. Instead of
requiring this profile, we can downgrade
the projects to 4.6.1 because they don't
need any specific feature from the new
version.

Also, to fix the build of the project
Nethereum.JsonRpc.IpcClient, we need to
define a conditional <PropertyGroup> that
defines the constant NET461, otherwise
the SendAsync() method is not found and
the build fails.