TcpClient: raise OperationCanceledException if cancelled
The approach of just stopping work (instead of throwing) could have
worked in the case of cancellation requested by the geewallet's
FaultTolerantParallelClient (because in this scenario, cancellation
means that the result has already been retreived, and the result
from the cancelled jobs will not be used).

But in the case that we have an external cancellationTokenSource,
not generated by the FaultTolerantParallelClient (e.g. generated
by an app-switch in a mobile platform), then we want to make sure
that the result from this computation is not used as a legitimate
result, and we do that by raising this exception.