Rpc(Http)Client: check if cancellation request after PostAsync()
We received a GatewayTimeout err with a delay greater than 1 second
after we had cancelled the task, and it happened in the method
EnsureSuccessStatusCode(). Given that the call before this method
is already sending the cancellationToken properly, I'm going to
inject a ThrowIfCancellationRequested call in between them, hoping
it fixes the problem(?).